Added Readme for Project

+++ main.py
@@ -200,8 +200,9 @@ def edit_post(post_id):
 @admin_only
 def delete_post(post_id):
     post_to_delete = db.get_or_404(BlogPost, post_id)
-    db.session.delete(post_to_delete)
-    db.session.commit()
+    if post_to_delete:
+        db.session.delete(post_to_delete)
+        db.session.commit()
     return redirect(url_for('get_all_posts'))
